school-portal-backend/requirements.txt
Brotli==1.0.9
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cloudinary==1.29.0
cssselect2==0.4.1
dnspython==2.1.0
firebase-admin==5.2.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-Passlib==0.1
Flask-WeasyPrint==0.6
fonttools==4.29.1
gcloud==0.18.3
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.45.0
grpcio-status==1.45.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jws==0.1.3
MarkupSafe==2.0.1
msgpack==1.0.3
oauth2client==4.1.3
passlib==1.6.1
pdfkit==1.0.0
Pillow==9.0.1
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pydyf==0.1.2
pymongo==4.0.1
pyparsing==3.0.7
pyphen==0.12.0
Pyrebase4==4.5.0
python-dotenv==0.19.2
python-jwt==2.0.1
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
tinycss2==1.1.1
uritemplate==4.1.1
urllib3==1.26.9
weasyprint==54.1
webencodings==0.5.1
Werkzeug==2.0.2
zopfli==0.1.9
